[VideoStrip] Add screenshot Lambda
Adds a new Lambda function that takes 20 equally spaced screenshots of a
video, creates a strip by joining the images side-by-side, and uploads
the result to an S3 bucket.

This function is triggered whenever a video gets uploaded to the input
bucket. The resulting strip is stored in the `videocloud-strips/` bucket.