[MultiRes] Resize video to different resolutions
When a user uploads a video, it triggers the `multiRes` function, which
will create resized copies of the video in 1080p, 720p, 540p, 360p and
144p, starting with the highest resolution that is equal or smaller than
the video's native resolution.

The resulting videos are uploaded to S3, in the `public/outputs/` folder
of the input bucket.

Also changes the strip creation function so that it uses the 144p
version of the video to create the strip.